init: Refactor & encapsulate registry code
Also use tracing in place of println!
