+++ crates/flakreate/src/registry.rs
@@ -1,18 +1,26 @@
+use std::{collections::BTreeMap, path::PathBuf};
+
 use glob::{Pattern, PatternError};
 use inquire::Select;
-use nix_rs::flake::url::FlakeUrl;
+use nix_rs::{
+    command::{NixCmd, NixCmdError},
+    flake::{eval::nix_eval_attr_json, url::FlakeUrl},
+};
+use thiserror::Error;
 
-use crate::flake_template::{self, template::FlakeTemplate};
+use crate::flake_template::{config::FlakeTemplateConfig, template::FlakeTemplate};
 
-pub struct FlakeTemplateRegistry {
+/// A [FlakeUrl] reference to a [TemplateRegistry]
+pub struct TemplateRegistryRef {
     pub flake_url: FlakeUrl,
+    /// A filter to apply to the templates. Unmatched templates will be ignored.
     pub filter: Option<Pattern>,
 }
 
-impl FlakeTemplateRegistry {
+impl TemplateRegistryRef {
     pub fn from_url(url: FlakeUrl) -> Result<Self, PatternError> {
         let (base, attr) = url.split_attr();
-        Ok(FlakeTemplateRegistry {
+        Ok(TemplateRegistryRef {
             flake_url: base,
             filter: if attr.is_none() {
                 None
@@ -27,6 +35,7 @@ impl FlakeTemplateRegistry {
         let templates = self.load_registry().await?;
         // TODO: avoid duplicates (aliases)
         let filtered_templates = templates
+            .0
             .iter()
             .filter(|template| {
                 self.filter
@@ -45,8 +54,93 @@ impl FlakeTemplateRegistry {
         Ok(template.clone())
     }
 
-    async fn load_registry(&self) -> anyhow::Result<Vec<FlakeTemplate>> {
-        let res = flake_template::template::fetch(&self.flake_url).await?;
+    async fn load_registry(&self) -> anyhow::Result<TemplateRegistry> {
+        let res = TemplateRegistry::from(&self.flake_url).await?;
         Ok(res)
     }
+}
+
+/// A registry of [FlakeTemplate]s
+pub struct TemplateRegistry(pub Vec<FlakeTemplate>);
+
+impl TemplateRegistry {
+    /// Fetch the templates defined in a flake
+    pub async fn from(url: &FlakeUrl) -> Result<Self, TemplateError> {
+        let v = if let Some(path) = url.as_local_path()
+            && let cache_file = path.join("flake.nix.json")
+            && cache_file.exists()
+        {
+            tracing::debug!("Fetching templates from cache: {}", cache_file.display());
+            Self::fetch_via_cache(cache_file).await?
+        } else {
+            tracing::debug!("Fetching templates from flake: {}", url);
+            Self::fetch_via_flake(url).await?
+        };
+        Ok(v)
+    }
+
+    async fn fetch_via_flake(url: &FlakeUrl) -> Result<Self, NixCmdError> {
+        let nixcmd = NixCmd::get().await;
+        let mut templates = nix_eval_attr_json::<BTreeMap<String, FlakeTemplate>>(
+            nixcmd,
+            &url.with_attr("templates"),
+        )
+        .await?
+        .unwrap_or_default();
+        let templates_config = nix_eval_attr_json::<BTreeMap<String, FlakeTemplateConfig>>(
+            nixcmd,
+            &url.with_attr("om.templates"),
+        )
+        .await?
+        .unwrap_or_default();
+        Self::set_template_deserialized_fields(&mut templates, &templates_config);
+        Ok(TemplateRegistry(templates.values().cloned().collect()))
+    }
+
+    /// Load from `flake.nix.json`
+    async fn fetch_via_cache(cache_file: PathBuf) -> Result<Self, CacheError> {
+        let flake_nix_json: serde_json::Value =
+            serde_json::from_reader(std::fs::File::open(cache_file)?)?;
+        let mut templates = serde_json::from_value(
+            flake_nix_json
+                .pointer("/templates")
+                .cloned()
+                .unwrap_or_default(),
+        )?;
+        let templates_config = serde_json::from_value(
+            flake_nix_json
+                .pointer("/om/templates")
+                .cloned()
+                .unwrap_or_default(),
+        )?;
+        Self::set_template_deserialized_fields(&mut templates, &templates_config);
+        Ok(TemplateRegistry(templates.values().cloned().collect()))
+    }
+
+    fn set_template_deserialized_fields(
+        templates: &mut BTreeMap<String, FlakeTemplate>,
+        templates_config: &BTreeMap<String, FlakeTemplateConfig>,
+    ) {
+        for (name, template) in templates.iter_mut() {
+            template.polyfill(name.clone(), templates_config.get(name).cloned());
+        }
+    }
+}
+
+#[derive(Error, Debug)]
+pub enum TemplateError {
+    #[error("Failed to fetch templates: {0}")]
+    FetchError(#[from] NixCmdError),
+
+    #[error("Failed to fetch templates from cache: {0}")]
+    CacheError(#[from] CacheError),
+}
+
+#[derive(Error, Debug)]
+pub enum CacheError {
+    #[error("Failed to read flake.nix.json: {0}")]
+    ReadError(#[from] std::io::Error),
+
+    #[error("Failed to parse flake.nix.json: {0}")]
+    ParseError(#[from] serde_json::Error),
 }
